final_app/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
psycopg[binary,pool]>=3.1.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.40
langgraph-checkpoint-postgres>=0.0.1

# Vector Store
qdrant-client>=1.7.0

# OpenAI
openai>=1.10.0

# Reranking (optional - for improved RAG relevance)
cohere>=5.0.0

# Document Processing
unstructured[all-docs]>=0.12.0
pypdf>=3.17.0
pdf2image>=1.16.0
pillow>=10.0.0

# AWS
boto3>=1.34.0

# Utilities
httpx>=0.26.0
tenacity>=8.2.3
python-dotenv>=1.0.0
numpy>=1.24.0

# Authentication
bcrypt>=4.0.0

# Rate Limiting
redis>=5.0.0
